Temporarily copy models/ai to the server module
Temporarily allow the server to work with models (the actual domain models) while I keep developing the initial versions of the game. The idea is to later quit this bad copy/paste approach and import a domain module to the game module and the server module. The domain models have to be decouple from the game models though (ebiten).