Reimplement graphic assets for Entity Diamond
The systems had to be more general and allow entities like `Path` that have 1:n relations, e.g. `Diamond` only has a `Graphic` but `Path` has one `Graphic` for when it's horizontal and other when it's vertical. This is also necessary for multiple sprites or animations of the same entity.